Move method `GetPostgresVesion` from doc store to diagnostics
