+++ src/Marten/Services/Diagnostics.cs
@@ -1,6 +1,8 @@
+using System;
 using Baseline;
 using Marten.Linq;
 using Marten.Schema;
+using Marten.Storage;
 using Marten.Util;
 using Npgsql;
 
@@ -9,6 +11,7 @@ namespace Marten.Services
     public class Diagnostics : IDiagnostics
     {
         private readonly DocumentStore _store;
+        private Version _postgreSqlVersion;
 
         public Diagnostics(DocumentStore store)
         {
@@ -47,6 +50,22 @@ namespace Marten.Services
             {
                 return conn.ExplainQuery(cmd);
             }
+        }
+
+        /// <summary>
+        /// Method to fetch Postgres server version
+        /// </summary>
+        /// <returns>Returns version</returns>
+        public Version GetPostgresVersion()
+        {
+            if (_postgreSqlVersion != null) return _postgreSqlVersion;
+
+            using (var conn = _store.Tenancy.Default.OpenConnection())
+            {
+                _postgreSqlVersion = conn.Connection.PostgreSqlVersion;
+            }
+
+            return _postgreSqlVersion;
         }
     }
 }
